23.11.23

+++ main.py
@@ -1,4 +1,6 @@
 import logging
+
+import asyncio
 import os
 
 from aiogram import Bot, Dispatcher, executor, types
@@ -12,37 +14,45 @@ logging.basicConfig(level=logging.INFO)
 bot = Bot(token=API_TOKEN, parse_mode='html')
 dp = Dispatcher(bot)
 
+'''@dp.message_handler()
+async def allll(message: types.Message):
+    print(message.from_user.id)'''
 
 
 @dp.message_handler(content_types="new_chat_members")
 async def on_user_join(message: types.Message):
-    print(message.chat.id)
-    await message.delete()
-    await bot.send_message(text='Bot rabotaet!))', chat_id= 404354012)
+    if message.chat.id == chat_id_admin:
+        print(message.chat.id)
+        await message.delete()
+        await bot.send_message(text='Bot rabotaet!))', chat_id= 404354012)
 
 @dp.message_handler(content_types="left_chat_member")
 async def on_user_join(message: types.Message):
-    print(message.chat.id)
-    await message.delete()
-    await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
+    if message.chat.id == chat_id_admin:
+        print(message.chat.id)
+        await message.delete()
+        await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
 
 @dp.message_handler(content_types="new_chat_title")
 async def on_user_join(message: types.Message):
-    print(message.chat.id)
-    await message.delete()
-    await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
+    if message.chat.id == chat_id_admin:
+        print(message.chat.id)
+        await message.delete()
+        await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
 
 @dp.message_handler(content_types="new_chat_photo")
 async def on_user_join(message: types.Message):
-    print(message.chat.id)
-    await message.delete()
-    await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
+    if message.chat.id == chat_id_admin:
+        print(message.chat.id)
+        await message.delete()
+        await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
 
 @dp.message_handler(content_types="delete_chat_photo")
 async def on_user_join(message: types.Message):
-    print(message.chat.id)
-    await message.delete()
-    await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
+    if message.chat.id == chat_id_admin:
+        print(message.chat.id)
+        await message.delete()
+        await bot.send_message(chat_id=404354012, text='Bot rabotaet!))')
 
 if __name__ == '__main__':
     executor.start_polling(dp, skip_updates=True)
